+++ R/add-variable.R
@@ -1,14 +1,66 @@
-addVariable <- function (dataset, vardef, ...) {
-    ## Construct payload (if not already constructed)
-    ## TODO: deprecate this behavior? (Only used in tests)
-    if (!inherits(vardef, "VariableDefinition")) {
-        vardef <- VariableDefinition(vardef, ...)
+##' Add multiple variables to a dataset
+##'
+##' This function lets you add more than one variable at a time to a dataset.
+##' If you have multiple variables to add, this function will be faster than
+##' doing \code{ds$var <- value} assignment because it doesn't refresh the
+##' dataset's state in between variable POST requests.
+##' @param dataset a CrunchDataset
+##' @param ... \code{\link{VariableDefinition}}s or a list of
+##' VariableDefinitions.
+##' @return \code{dataset} with the new variables added (invisibly)
+##' @export
+addVariables <- function (dataset, ...) {
+    var_catalog_url <- shojiURL(dataset, "catalogs", "variables")
+    
+    ## Get vardefs and validate
+    vardefs <- list(...)
+    ## Check for whether a list of vardefs passed
+    if (length(vardefs) == 1 && 
+        is.list(vardefs[[1]]) && 
+        !inherits(vardefs[[1]], "VariableDefinition")) {
+            
+        vardefs <- vardefs[[1]]
+    }
+    ## Check that all are VariableDefinitions
+    are.vardefs <- vapply(vardefs, inherits, logical(1),
+        what="VariableDefinition")
+    if (!all(are.vardefs)) {
+        halt("Must supply VariableDefinitions")
+    }
+    ## Check that if values specified, they have the right length
+    vardefs <- lapply(vardefs, validateVarDefRows, 
+        numrows=getNrow(dataset, filtered=FALSE))
+    
+    ## Upload one at a time.
+    ## TODO: Server should support bulk insert
+    new_var_urls <- lapply(vardefs, 
+        function (x) try(POSTNewVariable(var_catalog_url, x), silent=TRUE))
+        ## Be silent so we can throw the errors together at the end
+    
+    ## Check for errors
+    errs <- vapply(new_var_urls, is.error, logical(1))
+    if (any(errs)) {
+        if (length(errs) == 1) {
+            ## Just one variable added. Throw its error.
+            rethrow(new_var_urls[[1]])
+        }
+        halt("The following variable definition(s) errored on upload: ", 
+            paste(which(errs), collapse=", "), "\n", 
+            paste(unlist(lapply(new_var_urls[errs], errorMessage)), sep="\n"))
+        ## Could make better error message, return the URLs of the variables
+        ## that errored so that user can delete them and start over, etc.
     }
     
+    dataset <- refresh(dataset)
+    invisible(dataset)
+}
+
+validateVarDefRows <- function (vardef, numrows) {
+    ## Pre-check the column length being sent to the server to confirm that
+    ## the number of rows matches what's already in the dataset
     if (!any(c("expr", "subvariables") %in% names(vardef))) {
-        ## Validate that we're sending the right number of rows
         new <- length(vardef$values)
-        old <- getNrow(dataset, filtered=FALSE)
+        old <- numrows ## Just for naming clarity
         if (new == 1 && old > 1) {
             vardef$values <- rep(vardef$values, old)
             new <- old
@@ -19,10 +71,7 @@ addVariable <- function (dataset, vardef, ...) {
             halt("replacement has ", new, " rows, data has ", old)
         }
     }
-    var_url <- POSTNewVariable(shojiURL(dataset, "catalogs", "variables"), 
-        vardef)
-    dataset <- refresh(dataset)
-    invisible(dataset)
+    return(vardef)
 }
 
 POSTNewVariable <- function (catalog_url, variable) {
@@ -75,23 +124,4 @@ POSTNewVariable <- function (catalog_url, variable) {
     }
     out <- do.POST(variable)
     invisible(out)
-}
-
-addVariables <- function (dataset, vars) {
-    ## assume data frame
-    nvars <- ncol(vars)
-    vars_url <- variableCatalogURL(dataset)
-    for (i in seq_len(nvars)) {
-        POSTNewVariable(vars_url,
-            toVariable(vars[[i]], name=names(vars)[i], alias=names(vars)[i]))
-    }
-    invisible(refresh(dataset))
-}
-
-deriveVariable <- function (dataset, expr, ...) {
-    derivation <- list(...)
-    derivation$expr <- zcl(expr)
-    var_url <- POSTNewVariable(variableCatalogURL(dataset), derivation)
-    dataset <- refresh(dataset)
-    invisible(dataset)
 }