+++ R/categories.R
@@ -163,9 +163,6 @@ setMethod("is.na", "Categories", function (x) structure(vapply(x, is.na, logical
 
 n2i <- function (x, cats, strict=TRUE) {
     ## Convert x from category names to the corresponding category ids
-    if (is.variable(cats)) {
-        cats <- categories(cats)
-    }
     out <- ids(cats)[match(x, names(cats))]
     if (strict && any(is.na(out))) {
         halt(ifelse(sum(is.na(out)) > 1, "Categories", "Category"),