+++ R/expressions.R
@@ -8,7 +8,6 @@
 ##'
 ##' @param x an input
 ##' @param table For \code{\%in\%}. See \code{\link[base]{match}}
-##' @param mode For \code{as.vector}. Ignored.
 ##' @param resolution For \code{rollup}. Either \code{NULL} or a character in 
 ##' c("Y", "Q", "M", "W", "D", "h", "m", "s", "ms") indicating the unit of 
 ##' time at which a Datetime variable should be aggregated. If \code{NULL}, 
@@ -20,7 +19,7 @@
 ##' @name expressions
 NULL
 
-##' @rdname expressions
+##' @rdname variable-to-R
 ##' @export
 setMethod("as.vector", "CrunchExpr", function (x, mode) {
     payload <- list(command="select", variables=list(out=zcl(x)))
@@ -35,6 +34,11 @@ setMethod("as.vector", "CrunchExpr", function (x, mode) {
     return(columnParser(out$metadata$out$type, mode)(out$data$out, variable))
 })
 
+##' @rdname toVariable
+##' @export
+setMethod("toVariable", "CrunchExpr", function (x, ...) {
+    structure(list(expr=zcl(x), ...), class="VariableDefinition")
+})
 
 ## "Ops" for Crunch Variables
 ##